// enumeratewebapp/src/model/Listing2.ts
export interface Listing2 {
  TempKey: number;
  AREA_CODE: string;
  REG: string;
  REGN: string;
  CWT: string;
  CWTN: string;
  AMP: string;
  AMPN: string;
  TAM: string;
  TAMN: string;
  TYPE: string;
  MUN: string;
  MUNN: string;
  VIL: string;
  VILN: string;
  VILT: string;
  ST_A: string;
  ST_B: string;
  ST_C: string;
  ST_Time: string;
  FN_A: string;
  FN_B: string;
  FN_C: string;
  FN_Time: string;
  ENU_A: string;
  ENU_B: string;
  ENU_C: string;
  ACA_A: string;
  ACA_B: string;
  ACA_C: string;
  SUB_A: string;
  SUB_B: string;
  SUB_C: string;
  ENO: string;
  F1CODE: string;
  Eadd_A: string;
  Eadd_B: string;
  Eadd_C: string;
  Eadd_D: string;
  Ename_A: string;
  Ename_AT: string;
  Ename_B: string;
  Ename_C: string;
  Esex: string;
  EPID: string;
  EBD: string;
  ESI: string;
  status1: string;
  R01: string;
  R02: string;
  R03_A: string;
  R03_B: string;
  R03_C: string;
  lat: string;
  long: string;
  R07: string;
  R08: string;
  R08_sub: string;
  R09: string;
  R10: string;
  R11: string;
  R12: string;
  R12_sub: string;
  R13_A: string;
  R13_AT: string;
  R13_B: string;
  R13_C: string;
  R14: string;
  R15: string;
  R16_A: string;
  R16_B: string;
  R16_C: string;
  R17: string;
  R18: string;
  R19: string;
  R20: string;
  R21: string;
  R22: string;
  R23: string;
  R24: string;
  R25: string;
  R26: string;
  R27: string;
  Status2: string;
  R28: string;
  R29_A: string;
  R29_AT: string;
  R29_B: string;
  R29_C: string;
  R29_D: string;
  R29_E: string;
  R29_F: string;
  R29_G: string;
  R29_H: string;
  R29_I: string;
  R30: string;
  R31: string;
  R32: string;
  R33: string;
  R34: string;
  R35: string;
  R36: string;
  R37_A: string;
  R37_B: string;
  R37_C: string;
  R37_D: string;
  R37_E: string;
  R37_F: string;
  R37_G: string;
  ENU: string;
  AH_CODE: string; //Enumerate
  AH_CODE1: string; //TempListing
  FromE: string;
  Status: string;
  Active: string;
  IsTracking: string;
  TrackingStatus: string;
  IsMakeUp: string;
  SDATE_A: string;
  SDATE_B: string;
  SDATE_C: string;
  CDATE_A: string;
  CDATE_B: string;
  CDATE_C: string;
  P1_A: string;
  FROM_AREA_CODE: string;
  FROM_ADD: string;
}